+++ CountOnMe/Model/Calculator.swift
@@ -10,33 +10,43 @@ import Foundation
 
 class Calculator {
     // MARK: - Properties
+    //contains every number and operators
     var elements: [String] = []
 
-    var result: Float = 0
-    var total : Float = 0
+    //intermediate values for each operation
+    var result: Double = 0
+    var total : Double = 0
+
+    //final value of an operation
     var finalResult: String = "0"
+
+    //used to browse the array for the multiplicative group
     var leftIndex = 0
     var operatorIndex = 1
     var rightIndex = 2
 
+    //has the equal been pressed?
     var equalIsPressed: Bool = false
+    //is the operation still running?
     var isOperationStillRunning: Bool = false
+    //is the last index of elements an operator?
     var lastElementIsAnOperator: Bool = false
 
     // MARK: - Checkers
+    //checks if the expression is valid
     var expressionIsCorrect: Bool {
         return elements.last != "+" && elements.last != "-" && elements.last != "*" && elements.last != "/"
     }
-
+    //checks if there's enough elements to do an operation
     var expressionHaveEnoughElement: Bool {
         return elements.count >= 3
     }
-
+    //checks if an operator can be added
     var canAddOperator: Bool{
         continueOperation()
         return elements.last != "+" && elements.last != "-" && elements.last != "*" && elements.last != "/"
     }
-
+    //checks if the last item of elements is an operator
     var checkOperator: Bool {
         if elements.last == "+" || elements.last == "-" || elements.last == "*" || elements.last == "/" {
             return true
@@ -44,13 +54,13 @@ class Calculator {
             return false
         }
     }
-
+    //prevents any division by zero
     private func preventDivisionByZero() throws {
         if elements[operatorIndex] == "/" && elements[rightIndex] == "0" || elements[leftIndex] == "0" {
                 throw CalcError.divisionByZero
         }
     }
-
+    //error cases
     enum CalcError: Error {
         case notEnoughElements
         case expressionIsIncorrect
@@ -59,7 +69,8 @@ class Calculator {
 
 
     // MARK: - Functions
-    func addElement(number: Float) {
+    //adds a number to the array
+    func addElement(number: Double) {
         
         let elementsStringToNumber = number.clean
         checkIfNextInputIsANumberAfterOperationIsDone()
@@ -74,7 +85,7 @@ class Calculator {
             elements.append(finalValue)
         }
     }
-
+    //adds an operator to the array
     func addOperator(calcOperator: String) throws {
         if canAddOperator {
             elements.append(calcOperator)
@@ -82,14 +93,14 @@ class Calculator {
             throw CalcError.expressionIsIncorrect
         }
     }
-
+    //checks if elements contains only one element, and applies it's value to total
     private func defineTotalIfElementsContainsOneEntry() {
         if elements.count == 1 {
-            total = Float(elements[0])!
+            total = Double(elements[0])!
         }
     }
 
-    //Si le prochain input après un calcul est un chiffre, vider le tableau, sinon continuer
+    //checks if the next input after an operation is a number. Empties the array if so.
     private func checkIfNextInputIsANumberAfterOperationIsDone() {
         while equalIsPressed == true {
             if elements.last != "+" && elements.last != "-" && elements.last != "*" && elements.last != "/"{
@@ -99,7 +110,7 @@ class Calculator {
             }
         }
     }
-
+    //checks if the user continues to operate after pressing equal
     private func continueOperation() {
         while equalIsPressed == true {
             if elements.last != "+" && elements.last != "-" && elements.last != "*" && elements.last != "/"{
@@ -108,15 +119,16 @@ class Calculator {
             }
         }
     }
-
+    //processes all operations from additive group
     private func calcMultiplicativeGroup() throws {
         if expressionHaveEnoughElement {
             while rightIndex < elements.count {
                 var isNextOperatorFromMultiplicativeGroup = false
-                let left = Float(elements[leftIndex])!
+                let left = Double(elements[leftIndex])!
                 let operand = elements[operatorIndex]
-                let right = Float(elements[rightIndex])!
+                let right = Double(elements[rightIndex])!
 
+                //checks if a value is divided by zero
                 try preventDivisionByZero()
 
                 switch operand {
@@ -125,7 +137,6 @@ class Calculator {
                 default : isNextOperatorFromMultiplicativeGroup = true
                     break
                 }
-
                 if isNextOperatorFromMultiplicativeGroup {
                     leftIndex += 2
                     operatorIndex += 2
@@ -141,13 +152,13 @@ class Calculator {
             throw CalcError.notEnoughElements
         }
     }
-
+    //processes all operations from additive group
     private func calcAdditiveGroup() throws {
         if expressionHaveEnoughElement {
             while elements.count > 1 {
-                let left = Float(elements[0])!
+                let left = Double(elements[0])!
                 let operand = elements[1]
-                let right = Float(elements[2])!
+                let right = Double(elements[2])!
                 //var result: Float = 0
 
                 switch operand {
@@ -166,7 +177,7 @@ class Calculator {
             }
         }
     }
-
+    //when user taps equal button
     func tappedEqualButton() throws {
         try calcMultiplicativeGroup()
         defineTotalIfElementsContainsOneEntry()
@@ -177,7 +188,8 @@ class Calculator {
 }
 // MARK: - Extension
 
-extension Float {
+//truncates decimals if needed
+extension Double {
     var clean: String {
         return self.truncatingRemainder(dividingBy: 1) == 0 ? String(format: "%.0f", self) : String(self)
     }
